Cleanup and added cmc widget to footer
