+++ src/pages/developers/index.page.tsx
@@ -1,5 +1,5 @@
 import { SSRConfig, Trans, useTranslation } from 'next-i18next'
-import { Header } from '@components/commons/Header'
+import { PageHeader } from '@components/commons/PageHeader'
 import { Container } from '@components/commons/Container'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { DeveloperResourceSection } from './_components/DeveloperResourceSection'
@@ -16,12 +16,12 @@ export default function DevelopersPage (): JSX.Element {
         title={t('Head.title')}
         description={t('Head.desc')}
       />
-      <Header title={t('Header.title')}>
+      <PageHeader title={t('Header.title')}>
         <div className='mt-10 flex flex-wrap'>
           <div className='w-full text-2xl text-gray-900' data-testid='Header.desc.main'>{t('Header.desc')}
           </div>
         </div>
-      </Header>
+      </PageHeader>
 
       <Container>
         <Section title={t('DeveloperResourcesSection.title')} testId='DeveloperResourcesSection'>
