+++ src/pages/dex/index.page.tsx
@@ -1,5 +1,5 @@
 import { SSRConfig, useTranslation } from 'next-i18next'
-import { Header } from '@components/commons/Header'
+import { PageHeader } from '@components/commons/PageHeader'
 import { Container } from '@components/commons/Container'
 import { LiquidityPoolsSection } from './_components/LiquidityPoolsSection'
 import { WalletFeaturesSection } from './_components/WalletFeaturesSection'
@@ -15,7 +15,7 @@ export default function DexPage (): JSX.Element {
         title={t('Head.title')}
         description={t('Head.desc')}
       />
-      <Header title={t('Header.title')}>
+      <PageHeader title={t('Header.title')}>
         <div className='mt-10 flex flex-wrap'>
           <div className='w-full text-2xl text-gray-900' data-testid='Header.desc.main'>{t('Header.desc.main')}</div>
           <br />
@@ -25,7 +25,7 @@ export default function DexPage (): JSX.Element {
           >{t('Header.desc.secondary')}
           </div>
         </div>
-      </Header>
+      </PageHeader>
       <Container>
         <LiquidityPoolsSection />
         <WalletFeaturesSection />
