+++ src/pages/ecosystem/index.page.tsx
@@ -1,4 +1,4 @@
-import { Header } from '@components/commons/Header'
+import { PageHeader } from '@components/commons/PageHeader'
 import { Container } from '@components/commons/Container'
 import { SSRConfig, useTranslation } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
@@ -22,13 +22,13 @@ export default function EcosystemPage (): JSX.Element {
         title={t('Head.title')}
         description={t('Head.desc')}
       />
-      <Header title={t('Header.title')}>
+      <PageHeader title={t('Header.title')}>
         <div className='mt-10 flex flex-wrap'>
           <div className='w-full text-2xl text-gray-900' data-testid='Header.desc.main'>
             {t('Header.desc')}
           </div>
         </div>
-      </Header>
+      </PageHeader>
 
       <Container>
         <EcosystemSection title={t('ContributeSection.title')} testId='EcosystemSection.Contribute'>
