+++ src/pages/learn/index.page.tsx
@@ -1,4 +1,4 @@
-import { Header } from '@components/commons/Header'
+import { PageHeader } from '@components/commons/PageHeader'
 import { Container } from '@components/commons/Container'
 import { UserConfig, useTranslation } from 'next-i18next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
@@ -32,13 +32,13 @@ export default function LearnPage (props: InferGetStaticPropsType<typeof getStat
         title={t('Head.title')}
         description={t('Head.desc')}
       />
-      <Header title={t('Header.title')}>
+      <PageHeader title={t('Header.title')}>
         <div className='mt-10 flex flex-wrap'>
           <div className='w-full text-2xl text-gray-900' data-testid='Header.desc.main'>
             {t('Header.desc')}
           </div>
         </div>
-      </Header>
+      </PageHeader>
       <Container>
         <HowToSection posts={props.posts} />
         <FAQSection />
